Refactor auxiliary views
